Change how stoichiometric variables are handled
Previously had been using a single float, with a constraint
restricting its value to one of the permitted integers.
However, this doesn't work as iSAT only allows variables/expressions
to be raised to the power of integer constants, not variables.

New approach (mathcing what was previously done in the paper)
is to have a set of variables (one per possible value), of which
exactly one is 1 and the others are 0.